test(cli): split the CLI suite into parallelisable files
`pracht-cli.test.js` held 26 tests that each spawn the real CLI, and took
27.5s of the unit suite's 28.3s. Vitest parallelises across files but never
within one, and these tests block on execFileSync, so `it.concurrent` buys
nothing — only separate files reach separate workers.

Split it into nine `cli-*.test.js` files grouped by concern and balanced by
measured cost, sharing fixtures from `test/helpers/cli-fixtures.js`. Same
tests, same assertions; the only edit inside a test is collapsing eight
copies of the `execFileSync(tsc)` try/catch into `typecheckFixture()`.

Unit suite: 29.7s -> ~11s, with the test count unchanged at 1411.
