+++ packages/cli/test/cli-plan.test.js
@@ -1,0 +1,188 @@
+import { readFileSync } from "node:fs";
+import { join } from "node:path";
+
+import { afterEach, describe, expect, it } from "vitest";
+
+import {
+  cleanupTempDirs,
+  createRepoTempDir,
+  initializeGitRepo,
+  runCli,
+  runCliStatus,
+  writeInspectableManifestApp,
+  writeProjectFile,
+} from "./helpers/cli-fixtures.js";
+
+afterEach(cleanupTempDirs);
+
+describe("@pracht/cli plan and constraints", () => {
+  it("plans app-graph changes against a git base and enforces snapshot freshness", () => {
+    const appDir = createRepoTempDir("pracht-cli-plan-");
+    writeInspectableManifestApp(appDir);
+
+    runCli(["plan", "--write"], { cwd: appDir });
+    const snapshot = JSON.parse(readFileSync(join(appDir, ".pracht/app-graph.json"), "utf-8"));
+    expect(snapshot.routes.map((route) => route.path)).toEqual(["/dashboard"]);
+    expect(snapshot.api.map((route) => route.path)).toEqual(["/api/health", "/api/webhook"]);
+    // The agent-facing half of the graph is snapshotted too, so a change to it
+    // shows up in `pracht plan` like a route change does.
+    expect(snapshot.capabilities).toEqual([]);
+
+    initializeGitRepo(appDir);
+
+    // Add a route without refreshing the snapshot.
+    writeProjectFile(
+      appDir,
+      "src/routes.ts",
+      `import { defineApp, group, route, timeRevalidate } from "@pracht/core";
+
+export const app = defineApp({
+  shells: {
+    app: () => import("./shells/app.tsx"),
+  },
+  middleware: {
+    auth: () => import("./middleware/auth.ts"),
+  },
+  routes: [
+    group({ shell: "app", middleware: ["auth"] }, [
+      route("/dashboard", {
+        component: () => import("./routes/dashboard.tsx"),
+        loader: () => import("./server/dashboard-loader.ts"),
+        hydration: "full",
+        prefetch: "hover",
+        render: "isg",
+        revalidate: timeRevalidate(60),
+        speculation: "prefetch",
+      }),
+    ]),
+    route("/pricing", () => import("./routes/pricing.tsx"), { render: "ssg" }),
+  ],
+});
+`,
+    );
+    writeProjectFile(
+      appDir,
+      "src/routes/pricing.tsx",
+      `export function Component() {
+  return <main>Pricing</main>;
+}
+`,
+    );
+
+    const plan = JSON.parse(runCli(["plan", "--base", "HEAD", "--json"], { cwd: appDir }).stdout);
+    expect(plan.baseResolved).toBe("HEAD");
+    expect(plan.staleSnapshot).toBe(true);
+    expect(plan.diff.identical).toBe(false);
+    expect(plan.diff.addedRoutes.map((route) => route.path)).toEqual(["/pricing"]);
+    expect(plan.diff.changedRoutes).toEqual([]);
+    expect(plan.diff.removedRoutes).toEqual([]);
+
+    // The stale snapshot fails verification until it is rewritten.
+    const stale = runCliStatus(["verify", "--json"], { cwd: appDir });
+    expect(stale.status).toBe(1);
+    const staleReport = JSON.parse(stale.stdout);
+    expect(
+      staleReport.checks.some(
+        (check) =>
+          check.status === "error" &&
+          check.message.includes("snapshot") &&
+          check.message.includes("stale"),
+      ),
+    ).toBe(true);
+
+    runCli(["plan", "--write"], { cwd: appDir });
+    const fresh = JSON.parse(runCli(["verify", "--json"], { cwd: appDir }).stdout);
+    expect(
+      fresh.checks.some(
+        (check) =>
+          check.status === "ok" &&
+          check.message.includes("snapshot") &&
+          check.message.includes("up to date"),
+      ),
+    ).toBe(true);
+
+    const report = runCli(["report", "--base", "HEAD"], { cwd: appDir }).stdout;
+    expect(report).toContain("### App graph changes (base: `HEAD`)");
+    expect(report).toContain("+ route /pricing");
+    expect(report).toContain("### Verification");
+    expect(report).toContain("Generated by `pracht report`");
+  }, 120_000);
+
+  it("enforces declared app constraints in verify", () => {
+    const appDir = createRepoTempDir("pracht-cli-constraints-");
+    writeInspectableManifestApp(appDir);
+
+    writeProjectFile(
+      appDir,
+      "src/routes.ts",
+      `import { defineApp, requireHead, requireMiddleware, route } from "@pracht/core";
+
+export const app = defineApp({
+  shells: {
+    app: () => import("./shells/app.tsx"),
+  },
+  middleware: {
+    auth: () => import("./middleware/auth.ts"),
+  },
+  constraints: [requireMiddleware("/admin/**", "auth"), requireHead("/admin/**")],
+  routes: [route("/admin", () => import("./routes/admin.tsx"), { render: "ssr" })],
+});
+`,
+    );
+    writeProjectFile(
+      appDir,
+      "src/routes/admin.tsx",
+      `export function Component() {
+  return <main>Admin</main>;
+}
+`,
+    );
+
+    const failing = runCliStatus(["verify", "--json"], { cwd: appDir });
+    expect(failing.status).toBe(1);
+    const failingReport = JSON.parse(failing.stdout);
+    const errors = failingReport.checks.filter((check) => check.status === "error");
+    expect(
+      errors.some((check) => check.message.includes('missing required middleware "auth"')),
+    ).toBe(true);
+    expect(errors.some((check) => check.message.includes("does not export head()"))).toBe(true);
+
+    writeProjectFile(
+      appDir,
+      "src/routes.ts",
+      `import { defineApp, requireHead, requireMiddleware, route } from "@pracht/core";
+
+export const app = defineApp({
+  shells: {
+    app: () => import("./shells/app.tsx"),
+  },
+  middleware: {
+    auth: () => import("./middleware/auth.ts"),
+  },
+  constraints: [requireMiddleware("/admin/**", "auth"), requireHead("/admin/**")],
+  routes: [
+    route("/admin", () => import("./routes/admin.tsx"), { middleware: ["auth"], render: "ssr" }),
+  ],
+});
+`,
+    );
+    writeProjectFile(
+      appDir,
+      "src/routes/admin.tsx",
+      `export function head() {
+  return { title: "Admin" };
+}
+
+export function Component() {
+  return <main>Admin</main>;
+}
+`,
+    );
+
+    const passing = JSON.parse(runCli(["verify", "--json"], { cwd: appDir }).stdout);
+    expect(
+      passing.checks.some((check) => check.status === "ok" && check.message.includes("constraint")),
+    ).toBe(true);
+    expect(passing.checks.filter((check) => check.status === "error")).toEqual([]);
+  }, 120_000);
+});
