--- packages/cli/test/pracht-cli.test.js
@@ -1,2309 +1,0 @@
-import { execFileSync, spawn, spawnSync } from "node:child_process";
-import {
-  existsSync,
-  mkdirSync,
-  mkdtempSync,
-  readFileSync,
-  rmSync,
-  statSync,
-  utimesSync,
-  writeFileSync,
-} from "node:fs";
-import { dirname, join, resolve } from "node:path";
-import { tmpdir } from "node:os";
-import { fileURLToPath, pathToFileURL } from "node:url";
-
-import { afterEach, describe, expect, it } from "vitest";
-import ts from "typescript";
-
-const cliPath = fileURLToPath(new URL("../bin/pracht.js", import.meta.url));
-const repoRoot = resolve(dirname(cliPath), "../../..");
-const repoTempRoot = resolve(dirname(cliPath), "../test/.tmp");
-const coreImportPath = resolve(repoRoot, "packages/framework/src/index.ts");
-const capabilitiesImportPath = resolve(repoRoot, "packages/capabilities/src/index.ts");
-const nodeAdapterImportPath = resolve(repoRoot, "packages/adapter-node/src/index.ts");
-const vitePluginImportPath = resolve(repoRoot, "packages/vite-plugin/src/index.ts");
-const standardSchemaImportPath = resolve(
-  repoRoot,
-  "packages/framework/node_modules/@standard-schema/spec/dist/index.d.ts",
-);
-// The api-types compile test consumes the built declarations like a real
-// project would (skipLibCheck leaves them unchecked); compiling the framework
-// *source* alongside a populated Register augmentation is not supported.
-const coreDistTypesPath = resolve(repoRoot, "packages/framework/dist/index.d.mts");
-const capabilitiesDistTypesPath = resolve(repoRoot, "packages/capabilities/dist/index.d.mts");
-// Ambient declarations for `virtual:pracht/capabilities`, the module the Vite
-// plugin generates. The capability-types compile test includes it so the typed
-// browser client is checked exactly as an app would see it.
-const virtualTypesPath = resolve(repoRoot, "packages/vite-plugin/virtual.d.ts");
-const tscPath = resolve(repoRoot, "node_modules/typescript/bin/tsc");
-const tempDirs = [];
-
-afterEach(() => {
-  while (tempDirs.length > 0) {
-    rmSync(tempDirs.pop(), { force: true, recursive: true });
-  }
-});
-
-describe("@pracht/cli", () => {
-  it("reports the published package version", () => {
-    const packageJson = JSON.parse(
-      readFileSync(resolve(repoRoot, "packages/cli/package.json"), "utf-8"),
-    );
-
-    const env = { ...process.env };
-    delete env.NODE_ENV;
-    const result = spawnSync(process.execPath, [cliPath, "--version"], {
-      cwd: repoRoot,
-      encoding: "utf-8",
-      env,
-    });
-
-    expect(result.status).toBe(0);
-    expect(`${result.stdout}${result.stderr}`.trim()).toBe(packageJson.version);
-  });
-
-  it("scaffolds shell, middleware, route, and api modules for manifest apps", () => {
-    const appDir = createTempDir("pracht-cli-manifest-");
-    writeManifestApp(appDir);
-
-    runCli(["generate", "shell", "--name", "app"], { cwd: appDir });
-    runCli(["generate", "middleware", "--name", "auth"], { cwd: appDir });
-
-    const apiResult = runCli(
-      ["generate", "api", "--path", "/health", "--methods", "GET,POST", "--json"],
-      { cwd: appDir },
-    );
-    const apiJson = JSON.parse(apiResult.stdout);
-
-    runCli(
-      [
-        "generate",
-        "route",
-        "--path",
-        "/dashboard",
-        "--render",
-        "isg",
-        "--revalidate",
-        "120",
-        "--shell",
-        "app",
-        "--middleware",
-        "auth",
-        "--loader",
-        "--error-boundary",
-      ],
-      { cwd: appDir },
-    );
-
-    const manifest = readFileSync(join(appDir, "src/routes.ts"), "utf-8");
-    const shellSource = readFileSync(join(appDir, "src/shells/app.tsx"), "utf-8");
-    const middlewareSource = readFileSync(join(appDir, "src/middleware/auth.ts"), "utf-8");
-    const routeSource = readFileSync(join(appDir, "src/routes/dashboard.tsx"), "utf-8");
-    const apiSource = readFileSync(join(appDir, "src/api/health.ts"), "utf-8");
-
-    expect(apiJson).toMatchObject({
-      created: ["src/api/health.ts"],
-      kind: "api",
-      ok: true,
-      updated: [],
-    });
-    expect(shellSource).toContain("export function Shell({ children }: ShellProps)");
-    expect(middlewareSource).toContain("export const middleware: MiddlewareFn");
-    expect(routeSource).toContain("export async function loader(_args: LoaderArgs)");
-    expect(routeSource).toContain("export function ErrorBoundary({ error }: ErrorBoundaryProps)");
-    expect(apiSource).toContain("export function GET(_args: ApiRouteArgs)");
-    expect(apiSource).toContain("export async function POST({ request }: ApiRouteArgs)");
-    expect(manifest).toContain('import { defineApp, route, timeRevalidate } from "@pracht/core";');
-    expect(manifest).toContain('shells: {\n    app: "./shells/app.tsx",\n  },');
-    expect(manifest).toContain('middleware: {\n    auth: "./middleware/auth.ts",\n  },');
-    expect(manifest).toContain('route("/dashboard", "./routes/dashboard.tsx", {');
-    expect(manifest).toContain('shell: "app",');
-    expect(manifest).toContain('middleware: ["auth"],');
-    expect(manifest).toContain("revalidate: timeRevalidate(120)");
-  });
-
-  it("reports a healthy manifest app in doctor json output", () => {
-    const appDir = createTempDir("pracht-cli-doctor-ok-");
-    writeManifestApp(appDir, {
-      routesSource: `import { defineApp, route } from "@pracht/core";
-
-export const app = defineApp({
-  shells: {
-    app: "./shells/app.tsx",
-  },
-  middleware: {
-    auth: "./middleware/auth.ts",
-  },
-  routes: [route("/dashboard", "./routes/dashboard.tsx", { id: "dashboard", shell: "app", middleware: ["auth"], render: "ssr" })],
-});
-`,
-    });
-
-    writeProjectFile(
-      appDir,
-      "src/shells/app.tsx",
-      `import type { ShellProps } from "@pracht/core";
-
-export function Shell({ children }: ShellProps) {
-  return <main>{children}</main>;
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/middleware/auth.ts",
-      `import type { MiddlewareFn } from "@pracht/core";
-
-export const middleware: MiddlewareFn = async (_args, next) => {
-  return next();
-};
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/routes/dashboard.tsx",
-      `export function Component() {
-  return <h1>Dashboard</h1>;
-}
-`,
-    );
-
-    const result = runCli(["doctor", "--json"], { cwd: appDir });
-    const report = JSON.parse(result.stdout);
-
-    expect(report.ok).toBe(true);
-    expect(report.mode).toBe("manifest");
-    expect(report.checks.some((check) => check.message.includes("app manifest"))).toBe(true);
-    expect(report.checks.some((check) => check.message.includes("adapter dependency"))).toBe(true);
-  });
-
-  it("reports blocking doctor failures for broken manifest references", () => {
-    const appDir = createTempDir("pracht-cli-doctor-bad-");
-    writeManifestApp(appDir, {
-      routesSource: `import { defineApp, route } from "@pracht/core";
-
-export const app = defineApp({
-  routes: [route("/broken", "./routes/missing.tsx", { id: "broken", render: "ssr" })],
-});
-`,
-    });
-
-    const result = runCliStatus(["doctor", "--json"], { cwd: appDir });
-    expect(result.status).toBe(1);
-
-    const report = JSON.parse(result.stdout);
-    expect(report.ok).toBe(false);
-    expect(report.checks.some((check) => check.message.includes("missing files"))).toBe(true);
-  });
-
-  it.each(["404.tsx", "404/index.tsx"])(
-    "reports a pages-router %s as the not-found page instead of a route",
-    (notFoundFile) => {
-      const appDir = createTempDir("pracht-cli-doctor-pages-not-found-");
-      writePagesApp(appDir);
-      writeProjectFile(appDir, "src/pages/_app.tsx", "export function Shell() { return null; }");
-      writeProjectFile(
-        appDir,
-        "src/pages/index.tsx",
-        "export function Component() { return null; }",
-      );
-      writeProjectFile(
-        appDir,
-        `src/pages/${notFoundFile}`,
-        "export function Component() { return null; }",
-      );
-
-      const result = runCli(["doctor", "--json"], { cwd: appDir });
-      const report = JSON.parse(result.stdout);
-
-      expect(report.ok).toBe(true);
-      expect(report.checks.some((check) => check.message === "Found 1 page route.")).toBe(true);
-      expect(
-        report.checks.some((check) => check.message === "Found a pages-router not-found page."),
-      ).toBe(true);
-      expect(report.checks.some((check) => check.message === "Found 2 page routes.")).toBe(false);
-    },
-  );
-
-  it("rejects multiple pages-router not-found files", () => {
-    const appDir = createTempDir("pracht-cli-doctor-pages-not-found-duplicate-");
-    writePagesApp(appDir);
-    writeProjectFile(appDir, "src/pages/index.tsx", "export function Component() { return null; }");
-    writeProjectFile(appDir, "src/pages/404.tsx", "export function Component() { return null; }");
-    writeProjectFile(
-      appDir,
-      "src/pages/404/index.tsx",
-      "export function Component() { return null; }",
-    );
-
-    const result = runCliStatus(["doctor", "--json"], { cwd: appDir });
-    expect(result.status).toBe(1);
-
-    const report = JSON.parse(result.stdout);
-    expect(report.ok).toBe(false);
-    expect(report.checks.some((check) => check.message.includes("multiple not-found pages"))).toBe(
-      true,
-    );
-  });
-
-  it("rejects a second pages-router not-found file in changed verification", () => {
-    const appDir = createTempDir("pracht-cli-verify-pages-not-found-duplicate-");
-    writePagesApp(appDir);
-    writeProjectFile(appDir, "src/pages/index.tsx", "export function Component() { return null; }");
-    writeProjectFile(appDir, "src/pages/404.tsx", "export function Component() { return null; }");
-    initializeGitRepo(appDir);
-    writeProjectFile(
-      appDir,
-      "src/pages/404/index.tsx",
-      "export function Component() { return null; }",
-    );
-
-    const result = runCliStatus(["verify", "--changed", "--json"], { cwd: appDir });
-    expect(result.status).toBe(1);
-
-    const report = JSON.parse(result.stdout);
-    expect(report.ok).toBe(false);
-    expect(report.scope).toBe("changed");
-    expect(report.frameworkFiles).toContain("src/pages/404/index.tsx");
-    expect(report.checks.some((check) => check.message.includes("multiple not-found pages"))).toBe(
-      true,
-    );
-  });
-
-  it("reports a healthy manifest app in verify json output", () => {
-    const appDir = createTempDir("pracht-cli-verify-ok-");
-    writeManifestApp(appDir, {
-      routesSource: `import { defineApp, route } from "@pracht/core";
-
-export const app = defineApp({
-  routes: [route("/dashboard", "./routes/dashboard.tsx", { id: "dashboard", render: "ssr" })],
-});
-`,
-    });
-
-    writeProjectFile(
-      appDir,
-      "src/routes/dashboard.tsx",
-      `export function Component() {
-  return <h1>Dashboard</h1>;
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/api/health.ts",
-      `export function GET() {
-  return new Response("ok");
-}
-`,
-    );
-
-    const result = runCli(["verify", "--json"], { cwd: appDir });
-    const report = JSON.parse(result.stdout);
-
-    expect(report.ok).toBe(true);
-    expect(report.scope).toBe("full");
-    expect(report.checks.some((check) => check.message.includes("manifest module path"))).toBe(
-      true,
-    );
-    expect(report.checks.some((check) => check.message.includes("API route discovery"))).toBe(true);
-  });
-
-  it("reports duplicate API discovery failures in verify json output", () => {
-    const appDir = createTempDir("pracht-cli-verify-api-dupe-");
-    writeManifestApp(appDir);
-
-    writeProjectFile(
-      appDir,
-      "src/api/users.ts",
-      `export function GET() {
-  return new Response("users");
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/api/users/index.ts",
-      `export function GET() {
-  return new Response("users-index");
-}
-`,
-    );
-
-    const result = runCliStatus(["verify", "--json"], { cwd: appDir });
-    expect(result.status).toBe(1);
-
-    const report = JSON.parse(result.stdout);
-    expect(report.ok).toBe(false);
-    expect(report.checks.some((check) => check.message.includes("duplicate paths"))).toBe(true);
-  });
-
-  it("limits verify --changed to changed framework files", () => {
-    const appDir = createTempDir("pracht-cli-verify-changed-");
-    writeManifestApp(appDir, {
-      routesSource: `import { defineApp, route } from "@pracht/core";
-
-export const app = defineApp({
-  routes: [route("/dashboard", "./routes/dashboard.tsx", { id: "dashboard", render: "ssr" })],
-});
-`,
-    });
-
-    writeProjectFile(
-      appDir,
-      "src/routes/dashboard.tsx",
-      `export function Component() {
-  return <h1>Dashboard</h1>;
-}
-`,
-    );
-
-    initializeGitRepo(appDir);
-
-    writeProjectFile(
-      appDir,
-      "src/routes/dashboard.tsx",
-      `export function Component() {
-  return <h1>Updated dashboard</h1>;
-}
-`,
-    );
-    writeProjectFile(appDir, "notes.txt", "ignored");
-
-    const result = runCli(["verify", "--changed", "--json"], { cwd: appDir });
-    const report = JSON.parse(result.stdout);
-
-    expect(report.ok).toBe(true);
-    expect(report.scope).toBe("changed");
-    expect(report.frameworkFiles).toContain("src/routes/dashboard.tsx");
-    expect(report.frameworkFiles).not.toContain("notes.txt");
-    expect(
-      report.checks.some(
-        (check) =>
-          check.message.includes("Changed route module") &&
-          check.message.includes("src/routes/dashboard.tsx"),
-      ),
-    ).toBe(true);
-  });
-
-  it("inspects resolved routes, api handlers, and build metadata as JSON", () => {
-    const appDir = createRepoTempDir("pracht-cli-inspect-");
-    writeInspectableManifestApp(appDir);
-
-    const routes = JSON.parse(runCli(["inspect", "routes", "--json"], { cwd: appDir }).stdout);
-    const api = JSON.parse(runCli(["inspect", "api", "--json"], { cwd: appDir }).stdout);
-    const build = JSON.parse(runCli(["inspect", "build", "--json"], { cwd: appDir }).stdout);
-    const all = JSON.parse(runCli(["inspect", "--json"], { cwd: appDir }).stdout);
-
-    expect(routes).toEqual({
-      mode: "manifest",
-      notFound: null,
-      routes: [
-        {
-          file: "./routes/dashboard.tsx",
-          hydration: "full",
-          id: "dashboard",
-          loaderCache: null,
-          loaderFile: "./server/dashboard-loader.ts",
-          middleware: ["auth"],
-          path: "/dashboard",
-          prefetch: "hover",
-          render: "isg",
-          revalidate: {
-            kind: "time",
-            seconds: 60,
-          },
-          shell: "app",
-          shellFile: "./shells/app.tsx",
-          speculation: "prefetch",
-        },
-      ],
-    });
-
-    expect(api).toEqual({
-      api: [
-        {
-          file: "/src/api/health.ts",
-          hasDefaultHandler: false,
-          methods: ["GET", "POST"],
-          path: "/api/health",
-        },
-        {
-          file: "/src/api/webhook.ts",
-          hasDefaultHandler: true,
-          methods: [],
-          path: "/api/webhook",
-        },
-      ],
-      mode: "manifest",
-    });
-
-    expect(build).toEqual({
-      build: {
-        adapterTarget: "node",
-        clientEntryUrl: "/assets/client.js",
-        cssManifest: {
-          "src/routes/dashboard.tsx": ["/assets/dashboard.css"],
-          "src/shells/app.tsx": ["/assets/app.css"],
-        },
-        jsManifest: {
-          "src/routes/dashboard.tsx": ["/assets/dashboard.js", "/assets/vendor.js"],
-          "src/shells/app.tsx": ["/assets/app.js", "/assets/vendor.js"],
-          "virtual:pracht/client": ["/assets/vendor.js"],
-        },
-      },
-      mode: "manifest",
-    });
-
-    const capabilities = JSON.parse(
-      runCli(["inspect", "capabilities", "--json"], { cwd: appDir }).stdout,
-    );
-    expect(capabilities).toEqual({ capabilities: [], mode: "manifest" });
-
-    expect(all).toEqual({
-      ...routes,
-      ...api,
-      ...capabilities,
-      ...build,
-    });
-  }, 30_000);
-
-  it("plans app-graph changes against a git base and enforces snapshot freshness", () => {
-    const appDir = createRepoTempDir("pracht-cli-plan-");
-    writeInspectableManifestApp(appDir);
-
-    runCli(["plan", "--write"], { cwd: appDir });
-    const snapshot = JSON.parse(readFileSync(join(appDir, ".pracht/app-graph.json"), "utf-8"));
-    expect(snapshot.routes.map((route) => route.path)).toEqual(["/dashboard"]);
-    expect(snapshot.api.map((route) => route.path)).toEqual(["/api/health", "/api/webhook"]);
-    // The agent-facing half of the graph is snapshotted too, so a change to it
-    // shows up in `pracht plan` like a route change does.
-    expect(snapshot.capabilities).toEqual([]);
-
-    initializeGitRepo(appDir);
-
-    // Add a route without refreshing the snapshot.
-    writeProjectFile(
-      appDir,
-      "src/routes.ts",
-      `import { defineApp, group, route, timeRevalidate } from "@pracht/core";
-
-export const app = defineApp({
-  shells: {
-    app: () => import("./shells/app.tsx"),
-  },
-  middleware: {
-    auth: () => import("./middleware/auth.ts"),
-  },
-  routes: [
-    group({ shell: "app", middleware: ["auth"] }, [
-      route("/dashboard", {
-        component: () => import("./routes/dashboard.tsx"),
-        loader: () => import("./server/dashboard-loader.ts"),
-        hydration: "full",
-        prefetch: "hover",
-        render: "isg",
-        revalidate: timeRevalidate(60),
-        speculation: "prefetch",
-      }),
-    ]),
-    route("/pricing", () => import("./routes/pricing.tsx"), { render: "ssg" }),
-  ],
-});
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/routes/pricing.tsx",
-      `export function Component() {
-  return <main>Pricing</main>;
-}
-`,
-    );
-
-    const plan = JSON.parse(runCli(["plan", "--base", "HEAD", "--json"], { cwd: appDir }).stdout);
-    expect(plan.baseResolved).toBe("HEAD");
-    expect(plan.staleSnapshot).toBe(true);
-    expect(plan.diff.identical).toBe(false);
-    expect(plan.diff.addedRoutes.map((route) => route.path)).toEqual(["/pricing"]);
-    expect(plan.diff.changedRoutes).toEqual([]);
-    expect(plan.diff.removedRoutes).toEqual([]);
-
-    // The stale snapshot fails verification until it is rewritten.
-    const stale = runCliStatus(["verify", "--json"], { cwd: appDir });
-    expect(stale.status).toBe(1);
-    const staleReport = JSON.parse(stale.stdout);
-    expect(
-      staleReport.checks.some(
-        (check) =>
-          check.status === "error" &&
-          check.message.includes("snapshot") &&
-          check.message.includes("stale"),
-      ),
-    ).toBe(true);
-
-    runCli(["plan", "--write"], { cwd: appDir });
-    const fresh = JSON.parse(runCli(["verify", "--json"], { cwd: appDir }).stdout);
-    expect(
-      fresh.checks.some(
-        (check) =>
-          check.status === "ok" &&
-          check.message.includes("snapshot") &&
-          check.message.includes("up to date"),
-      ),
-    ).toBe(true);
-
-    const report = runCli(["report", "--base", "HEAD"], { cwd: appDir }).stdout;
-    expect(report).toContain("### App graph changes (base: `HEAD`)");
-    expect(report).toContain("+ route /pricing");
-    expect(report).toContain("### Verification");
-    expect(report).toContain("Generated by `pracht report`");
-  }, 120_000);
-
-  it("enforces declared app constraints in verify", () => {
-    const appDir = createRepoTempDir("pracht-cli-constraints-");
-    writeInspectableManifestApp(appDir);
-
-    writeProjectFile(
-      appDir,
-      "src/routes.ts",
-      `import { defineApp, requireHead, requireMiddleware, route } from "@pracht/core";
-
-export const app = defineApp({
-  shells: {
-    app: () => import("./shells/app.tsx"),
-  },
-  middleware: {
-    auth: () => import("./middleware/auth.ts"),
-  },
-  constraints: [requireMiddleware("/admin/**", "auth"), requireHead("/admin/**")],
-  routes: [route("/admin", () => import("./routes/admin.tsx"), { render: "ssr" })],
-});
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/routes/admin.tsx",
-      `export function Component() {
-  return <main>Admin</main>;
-}
-`,
-    );
-
-    const failing = runCliStatus(["verify", "--json"], { cwd: appDir });
-    expect(failing.status).toBe(1);
-    const failingReport = JSON.parse(failing.stdout);
-    const errors = failingReport.checks.filter((check) => check.status === "error");
-    expect(
-      errors.some((check) => check.message.includes('missing required middleware "auth"')),
-    ).toBe(true);
-    expect(errors.some((check) => check.message.includes("does not export head()"))).toBe(true);
-
-    writeProjectFile(
-      appDir,
-      "src/routes.ts",
-      `import { defineApp, requireHead, requireMiddleware, route } from "@pracht/core";
-
-export const app = defineApp({
-  shells: {
-    app: () => import("./shells/app.tsx"),
-  },
-  middleware: {
-    auth: () => import("./middleware/auth.ts"),
-  },
-  constraints: [requireMiddleware("/admin/**", "auth"), requireHead("/admin/**")],
-  routes: [
-    route("/admin", () => import("./routes/admin.tsx"), { middleware: ["auth"], render: "ssr" }),
-  ],
-});
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/routes/admin.tsx",
-      `export function head() {
-  return { title: "Admin" };
-}
-
-export function Component() {
-  return <main>Admin</main>;
-}
-`,
-    );
-
-    const passing = JSON.parse(runCli(["verify", "--json"], { cwd: appDir }).stdout);
-    expect(
-      passing.checks.some((check) => check.status === "ok" && check.message.includes("constraint")),
-    ).toBe(true);
-    expect(passing.checks.filter((check) => check.status === "error")).toEqual([]);
-  }, 120_000);
-
-  it("emits a Playwright smoke test alongside generated routes when e2e tooling exists", () => {
-    const appDir = createTempDir("pracht-cli-smoke-");
-    writeManifestApp(appDir);
-
-    // No Playwright setup: no test emitted.
-    runCli(["generate", "route", "--path", "/reports"], { cwd: appDir });
-    expect(existsSync(join(appDir, "e2e/reports.spec.ts"))).toBe(false);
-
-    // With an e2e directory the test is emitted by default.
-    writeProjectFile(appDir, "e2e/.gitkeep", "");
-    const result = JSON.parse(
-      runCli(["generate", "route", "--path", "/blog/:slug", "--json"], { cwd: appDir }).stdout,
-    );
-    expect(result.created).toContain("e2e/blog-slug.spec.ts");
-
-    const testSource = readFileSync(join(appDir, "e2e/blog-slug.spec.ts"), "utf-8");
-    expect(testSource).toContain('test("renders /blog/:slug"');
-    expect(testSource).toContain('page.goto("/blog/example-slug")');
-    expect(testSource).toContain('toHaveText("Slug")');
-
-    // --no-test opts out even when the setup exists.
-    runCli(["generate", "route", "--path", "/contact", "--no-test"], { cwd: appDir });
-    expect(existsSync(join(appDir, "e2e/contact.spec.ts"))).toBe(false);
-  });
-
-  it("generates typed route declarations and href helpers for manifest apps", () => {
-    const appDir = createRepoTempDir("pracht-cli-typegen-manifest-");
-    writeTypedManifestApp(appDir);
-
-    const result = JSON.parse(runCli(["typegen", "--json"], { cwd: appDir }).stdout);
-    const declaration = readFileSync(join(appDir, "src/pracht.d.ts"), "utf-8");
-    const runtime = readFileSync(join(appDir, "src/pracht-routes.ts"), "utf-8");
-
-    expect(result).toMatchObject({
-      apiRoutes: 3,
-      check: false,
-      files: ["src/pracht.d.ts", "src/pracht-routes.ts"],
-      mode: "manifest",
-      ok: true,
-      routes: 3,
-    });
-    expect(declaration).toContain(
-      'import type { ApiRouteMethodMap, RouteLoaderData, RouteParamInput, SearchParamsInput } from "@pracht/core";',
-    );
-    expect(declaration).toContain('"home": {');
-    expect(declaration).toContain("params: Record<never, never>;");
-    expect(declaration).toContain('"product": {');
-    expect(declaration).toContain('params: { "id": RouteParamInput; };');
-    // Route without a loader still points at its module; RouteLoaderData
-    // resolves to undefined until a loader export appears.
-    expect(declaration).toContain('data: RouteLoaderData<typeof import("./routes/home")>;');
-    // Inline loader.
-    expect(declaration).toContain('data: RouteLoaderData<typeof import("./routes/product")>;');
-    // Manifest-wired separate loader file wins over the route module.
-    expect(declaration).toContain(
-      'data: RouteLoaderData<typeof import("./server/dashboard-loader"), typeof import("./routes/dashboard")>;',
-    );
-    // API routes register on Register["apiRoutes"] for the typed apiFetch client.
-    expect(declaration).toContain("    apiRoutes: {");
-    expect(declaration).toContain('"/api/items/:id": {');
-    expect(declaration).toContain('methods: ApiRouteMethodMap<typeof import("./api/items/[id]")>;');
-    expect(runtime).toContain('id: "product"');
-    expect(runtime).toContain('path: "/products/:id"');
-    expect(runtime).toContain("export const href = createHref(routes);");
-    expect(existsSync(join(appDir, "api-module-loaded"))).toBe(false);
-
-    const check = JSON.parse(runCli(["typegen", "--check", "--json"], { cwd: appDir }).stdout);
-    expect(check).toMatchObject({ check: true, ok: true, routes: 3 });
-
-    // Unchanged outputs are not rewritten — dev-mode regeneration relies on
-    // this to avoid spurious HMR updates for the generated runtime module.
-    const declarationPath = join(appDir, "src/pracht.d.ts");
-    const past = new Date(Date.now() - 120_000);
-    utimesSync(declarationPath, past, past);
-    runCli(["typegen"], { cwd: appDir });
-    expect(statSync(declarationPath).mtimeMs).toBeLessThan(Date.now() - 60_000);
-
-    writeProjectFile(appDir, "src/pracht.d.ts", "stale\n");
-    const stale = runCliStatus(["typegen", "--check", "--json"], { cwd: appDir });
-    expect(stale.status).toBe(1);
-    expect(JSON.parse(stale.stderr)).toMatchObject({ ok: false });
-
-    for (const [declarationOut, runtimeOut] of [
-      ["src/collision.ts", "src/collision.ts"],
-      ["src/collision.d.ts", "src/collision.tsx"],
-    ]) {
-      const collision = runCliStatus(
-        ["typegen", "--out", declarationOut, "--runtime-out", runtimeOut, "--json"],
-        { cwd: appDir },
-      );
-      expect(collision.status).toBe(1);
-      expect(JSON.parse(collision.stderr)).toMatchObject({
-        ok: false,
-        error: expect.stringContaining("shares its basename"),
-      });
-    }
-  }, 30_000);
-
-  it("generates capability declarations from capability schemas", () => {
-    const appDir = createRepoTempDir("pracht-cli-typegen-capabilities-");
-    writeTypedManifestApp(appDir, { capabilities: true });
-
-    const result = JSON.parse(runCli(["typegen", "--json"], { cwd: appDir }).stdout);
-    const declaration = readFileSync(join(appDir, "src/pracht-capabilities.d.ts"), "utf-8");
-
-    expect(result).toMatchObject({
-      capabilities: 4,
-      files: ["src/pracht.d.ts", "src/pracht-routes.ts", "src/pracht-capabilities.d.ts"],
-      ok: true,
-    });
-    expect(declaration).toContain('declare module "@pracht/core"');
-    // Input: `limit` declares a default, so callers may omit it.
-    expect(declaration).toContain('"notes.search": {');
-    expect(declaration).toContain('input: { "query": string; "limit"?: number; };');
-    // Output: open objects keep an index signature; closed ones do not.
-    expect(declaration).toContain(
-      'output: { "notes": Array<Record<string, unknown>>; [key: string]: unknown; };',
-    );
-    expect(declaration).toContain('"notes.set-status": {');
-    expect(declaration).toContain('"status": "draft" | "published";');
-    expect(declaration).toContain('output: { "updated": true; };');
-
-    // The effect class and exposure drive the typed client: `destructive`
-    // requires a confirmation token and private capabilities are unreachable
-    // from the browser, so both have to reach the registration.
-    expect(declaration).toContain('effect: "read";');
-    expect(declaration).toContain('effect: "destructive";');
-    expect(declaration).toContain("exposed: { http: true; webmcp: false; mcp: false };");
-    // notes.set-status declares no `expose` — it stays private.
-    expect(declaration).toContain("exposed: { http: false; webmcp: false; mcp: false };");
-    // Contract prose becomes JSDoc so hovering a name shows what an agent reads.
-    expect(declaration).toContain("* Search notes");
-    expect(declaration).toContain("* Find notes matching a query.");
-    expect(declaration).toContain("capabilityClient: {");
-    expect(declaration).toContain('"notes": {');
-    expect(declaration).toContain('"search": CapabilityClientMethod<"notes.search">;');
-
-    const hoverPath = join(appDir, "src/capability-hover.ts");
-    writeProjectFile(
-      appDir,
-      "src/capability-hover.ts",
-      `import { capabilities } from "virtual:pracht/capabilities";
-capabilities.notes.search({ query: "roadmap" });
-`,
-    );
-    const hoverProgram = ts.createProgram(
-      [hoverPath, join(appDir, "src/pracht-capabilities.d.ts"), virtualTypesPath],
-      {
-        target: ts.ScriptTarget.ES2022,
-        module: ts.ModuleKind.ESNext,
-        moduleResolution: ts.ModuleResolutionKind.Bundler,
-        noEmit: true,
-        skipLibCheck: true,
-        strict: true,
-        paths: {
-          "@pracht/core": [coreDistTypesPath],
-          "@pracht/capabilities": [capabilitiesDistTypesPath],
-          "@standard-schema/spec": [standardSchemaImportPath],
-        },
-      },
-    );
-    const hoverSource = hoverProgram.getSourceFile(hoverPath);
-    const hoverChecker = hoverProgram.getTypeChecker();
-    let searchProperty;
-    const findSearchProperty = (node) => {
-      if (ts.isPropertyAccessExpression(node) && node.name.text === "search") {
-        searchProperty = node.name;
-      }
-      ts.forEachChild(node, findSearchProperty);
-    };
-    findSearchProperty(hoverSource);
-    const searchSymbol = hoverChecker.getSymbolAtLocation(searchProperty);
-    expect(ts.displayPartsToString(searchSymbol.getDocumentationComment(hoverChecker))).toBe(
-      "Search notes\n\nFind notes matching a query.",
-    );
-
-    const check = JSON.parse(runCli(["typegen", "--check", "--json"], { cwd: appDir }).stdout);
-    expect(check).toMatchObject({ capabilities: 4, check: true, ok: true });
-
-    // Removing every capability rewrites the existing file to the empty
-    // registration instead of leaving it stale.
-    writeTypedManifestApp(appDir, { capabilities: false });
-    const emptied = JSON.parse(runCli(["typegen", "--json"], { cwd: appDir }).stdout);
-    expect(emptied).toMatchObject({ capabilities: 0, ok: true });
-    expect(readFileSync(join(appDir, "src/pracht-capabilities.d.ts"), "utf-8")).toContain(
-      "capabilities: Record<never, never>;",
-    );
-  }, 30_000);
-
-  it("generates typed route declarations for pages-router apps", () => {
-    const appDir = createRepoTempDir("pracht-cli-typegen-pages-");
-    writeInspectablePagesApp(appDir);
-
-    const result = JSON.parse(runCli(["typegen", "--json"], { cwd: appDir }).stdout);
-    const declaration = readFileSync(join(appDir, "src/pracht.d.ts"), "utf-8");
-
-    expect(result).toMatchObject({ mode: "pages", ok: true, routes: 2 });
-    expect(declaration).toContain('"index": {');
-    expect(declaration).toContain('"blog-slug": {');
-    expect(declaration).toContain('params: { "slug": RouteParamInput; };');
-    expect(declaration).toContain('data: RouteLoaderData<typeof import("./pages/index")>;');
-    expect(declaration).toContain('data: RouteLoaderData<typeof import("./pages/blog/[slug]")>;');
-  }, 30_000);
-
-  it("generated api route types drive apiFetch end to end", () => {
-    const appDir = createRepoTempDir("pracht-cli-typegen-api-types-");
-    writeTypedManifestApp(appDir);
-    runCli(["typegen"], { cwd: appDir });
-
-    writeProjectFile(
-      appDir,
-      "src/api-consumer.ts",
-      `import { apiFetch, defineApi } from "@pracht/core";
-
-export async function main() {
-  // Params are required, convenient primitive inputs are stringified for the
-  // wire, and transformed schema output reaches the handler response type.
-  const item = await apiFetch("/api/items/:id", { params: { id: 1 } });
-  const _item: { id: number } = item;
-
-  // HEAD responses are always bodyless regardless of the handler return type.
-  const head = await apiFetch("/api/items/:id", { method: "HEAD", params: { id: "1" } });
-  const _head: undefined = head;
-
-  // @ts-expect-error - GET and HEAD requests cannot carry bodies
-  await apiFetch("/api/items/:id", { method: "HEAD", params: { id: "1" }, body: "nope" });
-
-  // @ts-expect-error - route params arrive as strings, so a number-only schema cannot validate
-  await apiFetch("/api/items/:id", { method: "DELETE", params: { id: 1 } });
-
-  // Body is typed by the route's Standard Schema input.
-  const created = await apiFetch("/api/items", { method: "POST", body: { name: "x" } });
-  const _created: { created: string } = created;
-
-  // @ts-expect-error - omitting method always means GET, never an inferred mutation
-  await apiFetch("/api/items", { body: { name: "x" } });
-
-  // A successful Response branch makes the parsed output unknowable.
-  const mixed = await apiFetch("/api/items", { method: "PUT", body: { id: 1 } });
-  // @ts-expect-error - mixed Response/JSON handlers have unknown output
-  const _mixed: { updated: boolean } = mixed;
-
-  const mutationMethod = Math.random() > 0.5 ? "POST" as const : "PUT" as const;
-  // @ts-expect-error - a union method must stay correlated with its body shape
-  await apiFetch("/api/items", { method: mutationMethod, body: { name: "x" } });
-
-  // @ts-expect-error - inferred JSON outputs must survive serialization unchanged
-  defineApi({ handler: () => ({ createdAt: new Date() }) });
-  // @ts-expect-error - undefined is not a top-level JSON value
-  defineApi({ handler: () => undefined });
-
-  // A default handler remains available for methods without a named override.
-  const fallback = await apiFetch("/api/items/:id", { method: "PATCH", params: { id: "1" } });
-  const _fallback: unknown = fallback;
-
-  // @ts-expect-error - unknown api path
-  await apiFetch("/api/nope");
-
-  // @ts-expect-error - missing required params
-  await apiFetch("/api/items/:id");
-
-  // @ts-expect-error - body must match the schema input
-  await apiFetch("/api/items", { method: "POST", body: { name: 1 } });
-
-  // @ts-expect-error - POST requires a body
-  await apiFetch("/api/items", { method: "POST" });
-
-  // @ts-expect-error - GET is not exported for /api/items
-  await apiFetch("/api/items", { method: "GET" });
-
-  // json() handlers keep their payload type (and status freedom) on the client.
-  const uploaded = await apiFetch("/api/uploads", {
-    method: "POST",
-    body: { name: "x", avatar: new File([], "a.png") },
-  });
-  const _uploaded: { uploaded: string } = uploaded;
-
-  // File-bearing body schemas accept FormData as the wire format.
-  await apiFetch("/api/uploads", { method: "POST", body: new FormData() });
-
-  // String query inputs stay fully typed.
-  const results = await apiFetch("/api/uploads", { query: { q: "boots" } });
-  const _results: { results: string[] } = results;
-
-  // @ts-expect-error - query values arrive as strings; a number input can never validate
-  await apiFetch("/api/uploads", { query: { q: "boots", page: 2 } });
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "tsconfig.json",
-      JSON.stringify(
-        {
-          compilerOptions: {
-            target: "ES2022",
-            module: "ESNext",
-            moduleResolution: "Bundler",
-            allowImportingTsExtensions: true,
-            noEmit: true,
-            strict: true,
-            skipLibCheck: true,
-            lib: ["ES2022", "DOM", "DOM.Iterable"],
-            jsx: "react-jsx",
-            jsxImportSource: "preact",
-            types: ["node", "vite/client"],
-            paths: {
-              "@pracht/core": [coreDistTypesPath],
-              "@standard-schema/spec": [standardSchemaImportPath],
-            },
-          },
-          include: ["src"],
-        },
-        null,
-        2,
-      ),
-    );
-
-    // Throws (non-zero exit) when the generated declaration fails to
-    // deliver end-to-end api types.
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", "."], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (error) {
-      throw new Error(error.stdout || error.stderr || error.message);
-    }
-  }, 60_000);
-
-  it("keeps the nested capability client usable before typegen", () => {
-    const appDir = createRepoTempDir("pracht-cli-untyped-capability-client-");
-
-    writeProjectFile(
-      appDir,
-      "src/capability-consumer.ts",
-      `import { capabilities } from "virtual:pracht/capabilities";
-
-export async function browser() {
-  const result = await capabilities.notes.search<{ notes: string[] }>({ query: "roadmap" });
-  if (result.ok) {
-    const _notes: string[] = result.data.notes;
-  }
-
-  const namespace = Math.random() > 0.5 ? "notes" : "projects";
-  await capabilities[namespace].search({ query: "dynamic" });
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "tsconfig.json",
-      JSON.stringify(
-        {
-          compilerOptions: {
-            target: "ES2022",
-            module: "ESNext",
-            moduleResolution: "Bundler",
-            noEmit: true,
-            strict: true,
-            skipLibCheck: true,
-            lib: ["ES2022", "DOM", "DOM.Iterable"],
-            paths: {
-              "@pracht/core": [coreDistTypesPath],
-              "@pracht/capabilities": [capabilitiesDistTypesPath],
-              "@standard-schema/spec": [standardSchemaImportPath],
-            },
-          },
-          files: [virtualTypesPath],
-          include: ["src"],
-        },
-        null,
-        2,
-      ),
-    );
-
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", "."], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (error) {
-      throw new Error(error.stdout || error.stderr || error.message);
-    }
-  }, 60_000);
-
-  it("generated capability types make the client APIs type-safe end to end", () => {
-    const appDir = createRepoTempDir("pracht-cli-typegen-capability-types-");
-    writeTypedManifestApp(appDir, { capabilities: true });
-    runCli(["typegen"], { cwd: appDir });
-
-    writeProjectFile(
-      appDir,
-      "src/capability-consumer.ts",
-      `import { createCapabilityTestHost, invokeCapability } from "@pracht/core";
-import type { CapabilityTestHost } from "@pracht/core";
-import { defineCapability } from "@pracht/capabilities";
-import type { CapabilityRunArgs } from "@pracht/capabilities";
-import { callCapability, capabilities, useCapability } from "virtual:pracht/capabilities";
-
-declare const ctx: { request: Request };
-declare const host: CapabilityTestHost;
-declare const readName: "notes.search" | "notes.stats";
-declare const mixedName: "notes.search" | "notes.stats";
-
-const fixtureHost = createCapabilityTestHost({
-  capabilities: {
-    "fixture.only": defineCapability({
-      title: "Fixture only",
-      description: "A capability that exists only in this standalone test host.",
-      input: {
-        type: "object",
-        properties: { value: { type: "string" } },
-        required: ["value"],
-        additionalProperties: false,
-      },
-      output: {
-        type: "object",
-        properties: { seen: { type: "string" } },
-        required: ["seen"],
-        additionalProperties: false,
-      },
-      effect: "read",
-      async run({ input }: CapabilityRunArgs<{ value: string }>) {
-        return { seen: input.value };
-      },
-    }),
-  },
-});
-
-export async function server() {
-  // A factory-created host reads both generics retained on the capability
-  // object. Typing run()'s input lets defineCapability infer its output.
-  const fixture = await fixtureHost.invoke("fixture.only", { value: "roadmap" });
-  if (fixture.ok) {
-    const _seen: string = fixture.data.seen;
-  }
-
-  // @ts-expect-error - the inferred standalone-host input rejects mismatches
-  await fixtureHost.invoke("fixture.only", { value: 42 });
-
-  // Input and output both come from the registration — no per-call generics.
-  const found = await invokeCapability("notes.search", { query: "roadmap" }, ctx);
-  if (found.ok) {
-    const _notes: Array<Record<string, unknown>> = found.data.notes;
-  }
-
-  // Private capabilities stay reachable from server code.
-  const updated = await invokeCapability("notes.set-status", { id: "1", status: "draft" }, ctx);
-  if (updated.ok) {
-    const _updated: true = updated.data.updated;
-  }
-
-  // @ts-expect-error - a mismatched input no longer falls through to an untyped overload
-  await invokeCapability("notes.search", { query: 42 }, ctx);
-
-  // @ts-expect-error - a missing required property is a compile error
-  await invokeCapability("notes.search", {}, ctx);
-
-  // @ts-expect-error - unknown capability name
-  await invokeCapability("notes.serach", { query: "x" }, ctx);
-
-  // @ts-expect-error - an explicit type argument cannot re-open the untyped form
-  await invokeCapability<{ notes: string[] }>("notes.search", { query: "x" }, ctx);
-
-  // @ts-expect-error - enums narrow to their literal members
-  await invokeCapability("notes.set-status", { id: "1", status: "archived" }, ctx);
-
-  const serverName = Math.random() > 0.5 ? "notes.search" as const : "notes.set-status" as const;
-  // @ts-expect-error - a union name cannot take input valid for only one possible capability
-  await invokeCapability(serverName, { query: "roadmap" }, ctx);
-}
-
-export async function browser() {
-  const found = await callCapability("notes.search", { query: "roadmap" });
-  if (found.ok) {
-    const _notes: Array<Record<string, unknown>> = found.data.notes;
-  }
-
-  // An empty input schema means the argument may be omitted entirely.
-  const stats = await callCapability("notes.stats");
-  if (stats.ok) {
-    const _total: number = stats.data.total;
-  }
-
-  // Destructive calls make their two phases explicit: prepare obtains the
-  // token without committing, then confirm commits the identical input.
-  await callCapability("notes.purge", { titlePrefix: "tmp" }, { prepare: true });
-  await callCapability("notes.purge", { titlePrefix: "tmp" }, { confirm: "token" });
-
-  // @ts-expect-error - destructive calls must explicitly prepare or confirm
-  await callCapability("notes.purge", { titlePrefix: "tmp" });
-
-  // @ts-expect-error - one call cannot prepare and commit simultaneously
-  await callCapability(
-    "notes.purge",
-    { titlePrefix: "tmp" },
-    { prepare: true, confirm: "token" },
-  );
-
-  // @ts-expect-error - notes.set-status is private: there is no endpoint to call
-  await callCapability("notes.set-status", { id: "1", status: "draft" });
-
-  // @ts-expect-error - mismatched input
-  await callCapability("notes.search", { query: 42 });
-
-  // @ts-expect-error - unknown capability name
-  await callCapability("notes.serach", { query: "x" });
-
-  const browserName = Math.random() > 0.5 ? "notes.search" as const : "notes.stats" as const;
-  // @ts-expect-error - stats may omit input, but search may not
-  await callCapability(browserName);
-  // @ts-expect-error - search input is not valid for every member of the name union
-  await callCapability(browserName, { query: "roadmap" });
-
-  const dynamic = useCapability(browserName);
-  // @ts-expect-error - the bound name must be narrowed before supplying member-specific input
-  await dynamic.call({ query: "roadmap" });
-}
-
-export async function client() {
-  // Dotted names become nested namespaces on the generated client.
-  const found = await capabilities.notes.search({ query: "roadmap" });
-  if (found.ok) {
-    const _notes: Array<Record<string, unknown>> = found.data.notes;
-  }
-
-  const stats = await capabilities.notes.stats();
-  if (stats.ok) {
-    const _total: number = stats.data.total;
-  }
-
-  await capabilities.notes.purge({ titlePrefix: "tmp" }, { prepare: true });
-  await capabilities.notes.purge({ titlePrefix: "tmp" }, { confirm: "token" });
-  await capabilities.notes.purge({ titlePrefix: "tmp" }, { prepare: true });
-
-  // @ts-expect-error - destructive calls must explicitly prepare or confirm
-  await capabilities.notes.purge({ titlePrefix: "tmp" });
-
-  const purge = useCapability("notes.purge");
-  await purge.call({ titlePrefix: "tmp" }, { prepare: true });
-  await purge.call({ titlePrefix: "tmp" }, { confirm: "token" });
-  // @ts-expect-error - the hook has the same destructive-call gate
-  await purge.call({ titlePrefix: "tmp" });
-
-  // @ts-expect-error - mismatched input
-  await capabilities.notes.search({ query: 42 });
-
-  // @ts-expect-error - private capabilities are absent from the browser client
-  await capabilities.notes["set-status"]({ id: "1", status: "draft" });
-
-  // @ts-expect-error - unknown capability
-  await capabilities.notes.nope({});
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "tsconfig.json",
-      JSON.stringify(
-        {
-          compilerOptions: {
-            target: "ES2022",
-            module: "ESNext",
-            moduleResolution: "Bundler",
-            allowImportingTsExtensions: true,
-            noEmit: true,
-            strict: true,
-            skipLibCheck: true,
-            lib: ["ES2022", "DOM", "DOM.Iterable"],
-            types: ["node"],
-            paths: {
-              "@pracht/core": [coreDistTypesPath],
-              "@pracht/capabilities": [capabilitiesDistTypesPath],
-              "@standard-schema/spec": [standardSchemaImportPath],
-            },
-          },
-          files: [virtualTypesPath],
-          include: ["src"],
-        },
-        null,
-        2,
-      ),
-    );
-
-    // Throws (non-zero exit) when any guarantee above stops holding — either a
-    // valid call starts failing or an invalid one stops being rejected.
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", "."], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (error) {
-      throw new Error(error.stdout || error.stderr || error.message);
-    }
-
-    // A current typegen file with no HTTP exposure is not a legacy file. Every
-    // capability must stay absent from all browser projections.
-    writeProjectFile(
-      appDir,
-      "src/routes.ts",
-      `import { defineApp } from "@pracht/core";
-
-export const app = defineApp({
-  capabilities: {
-    "notes.set-status": () => import("./capabilities/notes-set-status.ts"),
-  },
-  routes: [],
-});
-`,
-    );
-    runCli(["typegen"], { cwd: appDir });
-    writeProjectFile(
-      appDir,
-      "src/capability-consumer.ts",
-      `import { callCapability, capabilities, useCapability } from "virtual:pracht/capabilities";
-
-export async function browser() {
-  // @ts-expect-error - an all-private graph exposes no browser call names
-  await callCapability("notes.set-status", { id: "1", status: "draft" });
-  // @ts-expect-error - the nested client is empty for an all-private graph
-  await capabilities.notes["set-status"]({ id: "1", status: "draft" });
-  // @ts-expect-error - hooks cannot bind to private capabilities either
-  useCapability("notes.set-status");
-}
-`,
-    );
-
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", "."], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (error) {
-      throw new Error(error.stdout || error.stderr || error.message);
-    }
-
-    // Removing the last capability leaves an intentionally empty generated
-    // registration. That is still evidence typegen has run, so stale names
-    // must not regain the pre-typegen untyped fallback.
-    writeTypedManifestApp(appDir, { capabilities: false });
-    runCli(["typegen"], { cwd: appDir });
-    writeProjectFile(
-      appDir,
-      "src/capability-consumer.ts",
-      `import { invokeCapability } from "@pracht/core";
-import { callCapability, useCapability } from "virtual:pracht/capabilities";
-
-declare const ctx: { request: Request };
-
-export async function removed() {
-  // @ts-expect-error - an emitted empty graph exposes no stale browser names
-  await callCapability("notes.search", { query: "roadmap" });
-  // @ts-expect-error - hooks cannot bind to a removed capability
-  useCapability("notes.search");
-  // @ts-expect-error - direct invocation also knows the generated graph is empty
-  await invokeCapability("notes.search", { query: "roadmap" }, ctx);
-}
-`,
-    );
-
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", "."], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (error) {
-      throw new Error(error.stdout || error.stderr || error.message);
-    }
-
-    // A declaration generated before `effect`/`exposed` existed is committed in
-    // real apps, so it cannot be regenerated on the user's behalf. It must keep
-    // compiling exactly as it did: every registered name reachable from the
-    // browser, no confirmation gate, unknown names still rejected. This file is
-    // handwritten rather than generated precisely because no current typegen
-    // can emit it.
-    writeProjectFile(
-      appDir,
-      "src/pracht-capabilities.d.ts",
-      `import "@pracht/core";
-
-declare module "@pracht/core" {
-  interface Register {
-    capabilities: {
-      "notes.search": {
-        input: { "query": string; };
-        output: { "notes": string[]; };
-      };
-      "notes.purge": {
-        input: { "titlePrefix": string; };
-        output: { "purged": number; };
-      };
-      "notes.set-status": {
-        input: { "id": string; };
-        output: { "updated": true; };
-      };
-    };
-  }
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/capability-consumer.ts",
-      `import { invokeCapability } from "@pracht/core";
-import { callCapability, capabilities, useCapability } from "virtual:pracht/capabilities";
-
-declare const ctx: { request: Request };
-
-export async function legacyDeclaration() {
-  // Input and output types still apply — those fields have always been there.
-  const found = await callCapability("notes.search", { query: "roadmap" });
-  if (found.ok) {
-    const _notes: string[] = found.data.notes;
-  }
-  await invokeCapability("notes.search", { query: "roadmap" }, ctx);
-
-  // No \`effect\` recorded, so the confirmation gate cannot close. Demanding a
-  // token here would break every app upgrading with a committed declaration.
-  await callCapability("notes.purge", { titlePrefix: "tmp" });
-  const purge = useCapability("notes.purge");
-  await purge.call({ titlePrefix: "tmp" });
-
-  // No \`exposed\` recorded, so exposure goes unchecked and every registered
-  // name stays reachable — as it was before the field existed.
-  await callCapability("notes.set-status", { id: "1" });
-  await capabilities.notes["set-status"]({ id: "1" });
-
-  // @ts-expect-error - unknown names are rejected even by a legacy declaration
-  await callCapability("notes.serach", { query: "roadmap" });
-  // @ts-expect-error - and mismatched inputs still are too
-  await callCapability("notes.search", { query: 42 });
-}
-`,
-    );
-
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", "."], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (error) {
-      throw new Error(error.stdout || error.stderr || error.message);
-    }
-  }, 60_000);
-
-  it("names the valid capabilities when a call does not resolve", () => {
-    // Both lines below are compile errors under any of these designs, so only
-    // the message distinguishes them and only an assertion on the message keeps
-    // it from regressing. Gating the untyped form by resolving its `name` to
-    // `never` left that signature in place to absorb whatever arity filtering
-    // rejected, reporting every mistake as "not assignable to type 'never'";
-    // making one signature's arity depend on the name reported every mistake as
-    // an argument count. Splitting by effect class means an unresolved name is
-    // always arity-compatible with the first signature, which answers with the
-    // set of names that would have worked.
-    const appDir = createRepoTempDir("pracht-cli-typegen-capability-messages-");
-    writeTypedManifestApp(appDir, { capabilities: true });
-    runCli(["typegen"], { cwd: appDir });
-
-    writeProjectFile(
-      appDir,
-      "src/capability-consumer.ts",
-      `import { callCapability } from "virtual:pracht/capabilities";
-
-export async function mistakes() {
-  await callCapability("notes.serach", { query: "roadmap" });
-  await callCapability("notes.purge", { titlePrefix: "tmp" });
-}
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "tsconfig.json",
-      JSON.stringify(
-        {
-          compilerOptions: {
-            target: "ES2022",
-            module: "ESNext",
-            moduleResolution: "Bundler",
-            allowImportingTsExtensions: true,
-            noEmit: true,
-            strict: true,
-            skipLibCheck: true,
-            lib: ["ES2022", "DOM", "DOM.Iterable"],
-            types: ["node"],
-            paths: {
-              "@pracht/core": [coreDistTypesPath],
-              "@pracht/capabilities": [capabilitiesDistTypesPath],
-              "@standard-schema/spec": [standardSchemaImportPath],
-            },
-          },
-          files: [virtualTypesPath],
-          include: ["src"],
-        },
-        null,
-        2,
-      ),
-    );
-
-    let diagnostics = "";
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", ".", "--pretty", "false"], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-      throw new Error("expected the consumer to fail typechecking");
-    } catch (error) {
-      diagnostics = error.stdout || error.stderr || "";
-    }
-
-    // Nothing is reported against `never`, and nothing is reported as an
-    // argument count — both of those hide which name the compiler wanted.
-    expect(diagnostics).not.toContain("parameter of type 'never'");
-    expect(diagnostics).not.toContain("Expected 3 arguments, but got 2");
-    // The misspelled name is answered with the names that would have worked.
-    expect(diagnostics).toContain(
-      `error TS2345: Argument of type '"notes.serach"' is not assignable to parameter of type '"notes.search" | "notes.stats"'`,
-    );
-    // A destructive name is excluded from that same set, which is how "this
-    // one needs prepare/confirm" surfaces. Blunter than naming the rule, but it
-    // points at a real, listed distinction rather than at `never`.
-    expect(diagnostics).toContain(
-      `error TS2345: Argument of type '"notes.purge"' is not assignable to parameter of type '"notes.search" | "notes.stats"'`,
-    );
-  }, 60_000);
-
-  it("keeps all-private capability registrations out of browser client types", () => {
-    const appDir = createRepoTempDir("pracht-cli-typegen-private-capability-types-");
-    writeTypedManifestApp(appDir, { capabilities: true });
-
-    const manifestPath = join(appDir, "src/routes.ts");
-    const manifest = readFileSync(manifestPath, "utf-8").replace(
-      /  capabilities: \{[\s\S]*?\n  \},/,
-      `  capabilities: {
-    "notes.set-status": () => import("./capabilities/notes-set-status.ts"),
-  },`,
-    );
-    writeFileSync(manifestPath, manifest, "utf-8");
-    runCli(["typegen"], { cwd: appDir });
-
-    writeProjectFile(
-      appDir,
-      "src/private-capability-consumer.tsx",
-      `import { Form } from "@pracht/core";
-import { callCapability, capabilities } from "virtual:pracht/capabilities";
-
-// @ts-expect-error - the only registered capability has no HTTP endpoint
-callCapability("notes.set-status", { id: "1", status: "draft" });
-
-// @ts-expect-error - an all-private app has an empty nested browser client
-capabilities.notes["set-status"]({ id: "1", status: "draft" });
-
-// @ts-expect-error - capability forms also require an HTTP-exposed name
-const form = <Form capability="notes.set-status" />;
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "tsconfig.json",
-      JSON.stringify(
-        {
-          compilerOptions: {
-            target: "ES2022",
-            module: "ESNext",
-            moduleResolution: "Bundler",
-            allowImportingTsExtensions: true,
-            noEmit: true,
-            strict: true,
-            skipLibCheck: true,
-            lib: ["ES2022", "DOM", "DOM.Iterable"],
-            jsx: "react-jsx",
-            jsxImportSource: "preact",
-            types: ["node"],
-            paths: {
-              "@pracht/core": [coreDistTypesPath],
-              "@pracht/capabilities": [capabilitiesDistTypesPath],
-              "@standard-schema/spec": [standardSchemaImportPath],
-            },
-          },
-          files: [virtualTypesPath],
-          include: ["src"],
-        },
-        null,
-        2,
-      ),
-    );
-
-    try {
-      execFileSync(process.execPath, [tscPath, "-p", "."], {
-        cwd: appDir,
-        encoding: "utf-8",
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (error) {
-      throw new Error(error.stdout || error.stderr || error.message);
-    }
-  }, 60_000);
-
-  it("pracht dev explains how to enable generated route types", async () => {
-    const appDir = createRepoTempDir("pracht-cli-dev-typegen-hint-");
-    writeTypedManifestApp(appDir);
-
-    const child = spawn(process.execPath, [cliPath, "dev"], {
-      cwd: appDir,
-      env: process.env,
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-    let output = "";
-    child.stdout.setEncoding("utf-8");
-    child.stderr.setEncoding("utf-8");
-    child.stdout.on("data", (chunk) => {
-      output += chunk;
-    });
-    child.stderr.on("data", (chunk) => {
-      output += chunk;
-    });
-
-    try {
-      await waitFor(
-        () => output.includes("run `pracht typegen` once"),
-        30_000,
-        () => output,
-      );
-      expect(existsSync(join(appDir, "src/pracht.d.ts"))).toBe(false);
-    } finally {
-      child.kill("SIGTERM");
-    }
-  }, 90_000);
-
-  it("pracht dev keeps generated route types in sync with route files", async () => {
-    const appDir = createRepoTempDir("pracht-cli-dev-typegen-");
-    writeTypedManifestApp(appDir);
-    writeProjectFile(
-      appDir,
-      "src/routes.ts",
-      `import { defineApp } from "@pracht/core";
-import { routes } from "./route-definitions";
-
-export const app = defineApp({ routes });
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/route-definitions.ts",
-      `import { route } from "@pracht/core";
-
-export const routes = [
-  route("/", "./routes/home.tsx", { id: "home", render: "ssg" }),
-];
-`,
-    );
-    runCli(["typegen"], { cwd: appDir });
-
-    const child = spawn(process.execPath, [cliPath, "dev"], {
-      cwd: appDir,
-      env: process.env,
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-    let output = "";
-    child.stdout.setEncoding("utf-8");
-    child.stderr.setEncoding("utf-8");
-    child.stdout.on("data", (chunk) => {
-      output += chunk;
-    });
-    child.stderr.on("data", (chunk) => {
-      output += chunk;
-    });
-
-    try {
-      // The route-type watcher attaches before the banner prints.
-      await waitFor(
-        () => output.includes("http"),
-        30_000,
-        () => output,
-      );
-
-      writeProjectFile(
-        appDir,
-        "src/api/ping.ts",
-        "export function GET() {\n  return Response.json({ pong: true });\n}\n",
-      );
-
-      await waitFor(
-        () => readFileSync(join(appDir, "src/pracht.d.ts"), "utf-8").includes('"/api/ping"'),
-        30_000,
-        () => output,
-      );
-
-      writeProjectFile(
-        appDir,
-        "src/route-definitions.ts",
-        `import { route } from "@pracht/core";
-
-export const routes = [
-  route("/", "./routes/home.tsx", { id: "home", render: "ssg" }),
-  route("/settings", "./routes/home.tsx", { id: "settings", render: "ssr" }),
-];
-`,
-      );
-
-      await waitFor(
-        () => readFileSync(join(appDir, "src/pracht.d.ts"), "utf-8").includes('"settings"'),
-        30_000,
-        () => output,
-      );
-    } finally {
-      child.kill("SIGTERM");
-    }
-  }, 90_000);
-
-  it("scaffolds pages-router routes without touching a manifest", () => {
-    const appDir = createTempDir("pracht-cli-pages-");
-    writePagesApp(appDir);
-
-    runCli(["generate", "route", "--path", "/blog/:slug", "--render", "ssg", "--loader"], {
-      cwd: appDir,
-    });
-
-    const routePath = join(appDir, "src/pages/blog/[slug].tsx");
-    expect(existsSync(routePath)).toBe(true);
-
-    const routeSource = readFileSync(routePath, "utf-8");
-    expect(routeSource).toContain('export const RENDER_MODE = "ssg";');
-    expect(routeSource).toContain("export function getStaticPaths()");
-    expect(routeSource).toContain('slug: "example-slug"');
-  });
-});
-
-async function waitFor(predicate, timeoutMs, describeFailure) {
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() < deadline) {
-    if (predicate()) {
-      return;
-    }
-    await new Promise((resolve) => setTimeout(resolve, 250));
-  }
-  throw new Error(
-    `Timed out waiting for condition.${describeFailure ? `\n${describeFailure()}` : ""}`,
-  );
-}
-
-function createTempDir(prefix) {
-  const dir = mkdtempSync(join(tmpdir(), prefix));
-  tempDirs.push(dir);
-  return dir;
-}
-
-function createRepoTempDir(prefix) {
-  mkdirSync(repoTempRoot, { recursive: true });
-  const dir = mkdtempSync(join(repoTempRoot, prefix));
-  tempDirs.push(dir);
-  return dir;
-}
-
-function runCli(args, { cwd }) {
-  return {
-    stdout: execFileSync(process.execPath, [cliPath, ...args], {
-      cwd,
-      encoding: "utf-8",
-      env: process.env,
-      stdio: ["ignore", "pipe", "pipe"],
-    }),
-  };
-}
-
-function runCliStatus(args, { cwd }) {
-  return spawnSync(process.execPath, [cliPath, ...args], {
-    cwd,
-    encoding: "utf-8",
-    env: process.env,
-  });
-}
-
-function initializeGitRepo(appDir) {
-  execFileSync("git", ["init"], {
-    cwd: appDir,
-    env: process.env,
-    stdio: "ignore",
-  });
-  execFileSync("git", ["config", "user.email", "test@example.com"], {
-    cwd: appDir,
-    env: process.env,
-    stdio: "ignore",
-  });
-  execFileSync("git", ["config", "user.name", "Pracht Tests"], {
-    cwd: appDir,
-    env: process.env,
-    stdio: "ignore",
-  });
-  execFileSync("git", ["add", "."], {
-    cwd: appDir,
-    env: process.env,
-    stdio: "ignore",
-  });
-  execFileSync("git", ["commit", "-m", "initial"], {
-    cwd: appDir,
-    env: process.env,
-    stdio: "ignore",
-  });
-}
-
-function writeManifestApp(appDir, { routesSource } = {}) {
-  writeProjectFile(
-    appDir,
-    "package.json",
-    JSON.stringify(
-      {
-        name: "fixture-app",
-        private: true,
-        dependencies: {
-          "@pracht/adapter-node": "workspace:*",
-          "@pracht/cli": "workspace:*",
-        },
-      },
-      null,
-      2,
-    ),
-  );
-  writeProjectFile(
-    appDir,
-    "vite.config.ts",
-    `import { defineConfig } from "vite";
-import { pracht } from "@pracht/vite-plugin";
-
-export default defineConfig({
-  plugins: [pracht()],
-});
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/routes.ts",
-    routesSource ??
-      `import { defineApp } from "@pracht/core";
-
-export const app = defineApp({
-  routes: [],
-});
-`,
-  );
-}
-
-function writeTypedManifestApp(appDir, { capabilities = false } = {}) {
-  const vitePluginImport = pathToFileURL(vitePluginImportPath).href;
-
-  writeProjectFile(
-    appDir,
-    "package.json",
-    JSON.stringify(
-      {
-        name: "fixture-typegen-app",
-        private: true,
-        type: "module",
-      },
-      null,
-      2,
-    ),
-  );
-  writeProjectFile(
-    appDir,
-    "vite.config.ts",
-    `import { defineConfig } from "vite";
-import { pracht } from ${JSON.stringify(vitePluginImport)};
-
-export default defineConfig({
-  plugins: [pracht()],
-  resolve: {
-    alias: {
-      "@pracht/adapter-node": ${JSON.stringify(nodeAdapterImportPath)},
-      "@pracht/capabilities": ${JSON.stringify(capabilitiesImportPath)},
-      "@pracht/core": ${JSON.stringify(coreImportPath)},
-    },
-  },
-});
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/routes.ts",
-    `import { defineApp, route } from "@pracht/core";
-
-export const app = defineApp({
-${
-  capabilities
-    ? `  capabilities: {
-    "notes.search": () => import("./capabilities/notes-search.ts"),
-    "notes.set-status": () => import("./capabilities/notes-set-status.ts"),
-    "notes.purge": () => import("./capabilities/notes-purge.ts"),
-    "notes.stats": () => import("./capabilities/notes-stats.ts"),
-  },
-`
-    : ""
-}  routes: [
-    route("/", "./routes/home.tsx", { id: "home", render: "ssg" }),
-    route("/products/:id", "./routes/product.tsx", { id: "product", render: "ssr" }),
-    route("/dashboard", {
-      component: "./routes/dashboard.tsx",
-      loader: "./server/dashboard-loader.ts",
-      id: "dashboard",
-      render: "ssr",
-    }),
-  ],
-});
-`,
-  );
-  if (capabilities) {
-    writeProjectFile(
-      appDir,
-      "src/capabilities/notes-search.ts",
-      `import { defineCapability } from "@pracht/capabilities";
-
-export default defineCapability({
-  title: "Search notes",
-  description: "Find notes matching a query.",
-  input: {
-    type: "object",
-    properties: {
-      query: { type: "string", minLength: 1 },
-      limit: { type: "integer", minimum: 1, maximum: 20, default: 10 },
-    },
-    required: ["query"],
-    additionalProperties: false,
-  },
-  output: {
-    type: "object",
-    properties: { notes: { type: "array", items: { type: "object" } } },
-    required: ["notes"],
-  },
-  effect: "read",
-  expose: { http: true },
-  async run() {
-    return { notes: [] };
-  },
-});
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/capabilities/notes-set-status.ts",
-      `import { defineCapability } from "@pracht/capabilities";
-
-export default defineCapability({
-  title: "Set note status",
-  description: "Move a note between draft and published.",
-  input: {
-    type: "object",
-    properties: {
-      id: { type: "string" },
-      status: { enum: ["draft", "published"] },
-    },
-    required: ["id", "status"],
-    additionalProperties: false,
-  },
-  output: {
-    type: "object",
-    properties: { updated: { const: true } },
-    required: ["updated"],
-    additionalProperties: false,
-  },
-  effect: "write",
-  async run() {
-    return { updated: true };
-  },
-});
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/capabilities/notes-purge.ts",
-      `import { defineCapability } from "@pracht/capabilities";
-
-export default defineCapability({
-  title: "Purge notes",
-  description: "Permanently delete notes matching a prefix.",
-  input: {
-    type: "object",
-    properties: { titlePrefix: { type: "string" } },
-    required: ["titlePrefix"],
-    additionalProperties: false,
-  },
-  output: {
-    type: "object",
-    properties: { purged: { type: "integer" } },
-    required: ["purged"],
-    additionalProperties: false,
-  },
-  effect: "destructive",
-  expose: { http: true },
-  async run() {
-    return { purged: 0 };
-  },
-});
-`,
-    );
-    writeProjectFile(
-      appDir,
-      "src/capabilities/notes-stats.ts",
-      `import { defineCapability } from "@pracht/capabilities";
-
-export default defineCapability({
-  title: "Note stats",
-  description: "Counts across every note.",
-  input: { type: "object", properties: {}, additionalProperties: false },
-  output: {
-    type: "object",
-    properties: { total: { type: "integer" } },
-    required: ["total"],
-    additionalProperties: false,
-  },
-  effect: "read",
-  expose: { http: true },
-  async run() {
-    return { total: 0 };
-  },
-});
-`,
-    );
-  }
-  writeProjectFile(appDir, "src/routes/home.tsx", "export function Component() { return null; }\n");
-  writeProjectFile(
-    appDir,
-    "src/routes/product.tsx",
-    `export async function loader() {
-  return { product: { id: "sku-1" } };
-}
-
-export function Component() { return null; }
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/routes/dashboard.tsx",
-    "export function Component() { return null; }\n",
-  );
-  writeProjectFile(
-    appDir,
-    "src/server/dashboard-loader.ts",
-    `export async function loader() {
-  return { widgets: 3 };
-}
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/lib/schema-util.ts",
-    `import type { StandardSchemaV1 } from "@standard-schema/spec";
-
-export function passthroughSchema<TInput, TOutput = TInput>(): StandardSchemaV1<TInput, TOutput> {
-  return {
-    "~standard": {
-      version: 1,
-      vendor: "fixture",
-      validate: (value) => ({ value: value as TOutput }),
-    },
-  };
-}
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/api/items/[id].ts",
-    `import { defineApi } from "@pracht/core";
-import { passthroughSchema } from "../../lib/schema-util";
-
-export const GET = defineApi({
-  params: passthroughSchema<{ id: string }, { id: number }>(),
-  handler: ({ params }) => ({ id: params.id }),
-});
-
-export const HEAD = defineApi({
-  handler: ({ params }) => ({ id: params.id }),
-});
-
-export default async function handler() {
-  return new Response("fallback");
-}
-
-export const DELETE = defineApi({
-  params: passthroughSchema<{ id: number }>(),
-  handler: () => ({ deleted: true }),
-});
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/api/items/index.ts",
-    `import { writeFileSync } from "node:fs";
-import { defineApi } from "@pracht/core";
-import { passthroughSchema } from "../../lib/schema-util";
-
-writeFileSync("api-module-loaded", "typegen executed an API module");
-
-export const POST = defineApi({
-  body: passthroughSchema<{ name: string }>(),
-  handler: ({ body }) => ({ created: body.name }),
-});
-
-export const PUT = defineApi({
-  body: passthroughSchema<{ id: number }>(),
-  handler: () =>
-    Math.random() > 0.5 ? { updated: true } : new Response(null, { status: 204 }),
-});
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/api/uploads.ts",
-    `import { defineApi, json } from "@pracht/core";
-import { passthroughSchema } from "../lib/schema-util";
-
-export const POST = defineApi({
-  body: passthroughSchema<{ name: string; avatar: File }>(),
-  handler: ({ body }) => json({ uploaded: body.name }, { status: 201 }),
-});
-
-export const GET = defineApi({
-  query: passthroughSchema<{ q: string; page?: number }>(),
-  handler: ({ query }) => ({ results: [query.q] }),
-});
-`,
-  );
-}
-
-function writeInspectablePagesApp(appDir) {
-  const vitePluginImport = pathToFileURL(vitePluginImportPath).href;
-
-  writeProjectFile(
-    appDir,
-    "package.json",
-    JSON.stringify(
-      {
-        name: "fixture-typegen-pages-app",
-        private: true,
-        type: "module",
-      },
-      null,
-      2,
-    ),
-  );
-  writeProjectFile(
-    appDir,
-    "vite.config.ts",
-    `import { defineConfig } from "vite";
-import { pracht } from ${JSON.stringify(vitePluginImport)};
-
-export default defineConfig({
-  plugins: [pracht({ pagesDir: "/src/pages" })],
-  resolve: {
-    alias: {
-      "@pracht/adapter-node": ${JSON.stringify(nodeAdapterImportPath)},
-      "@pracht/core": ${JSON.stringify(coreImportPath)},
-    },
-  },
-});
-`,
-  );
-  writeProjectFile(appDir, "src/pages/index.tsx", "export function Component() { return null; }\n");
-  writeProjectFile(
-    appDir,
-    "src/pages/blog/[slug].tsx",
-    `export async function loader() {
-  return { slug: "hello" };
-}
-
-export function Component() { return null; }
-`,
-  );
-}
-
-function writeInspectableManifestApp(appDir) {
-  const vitePluginImport = pathToFileURL(vitePluginImportPath).href;
-
-  writeProjectFile(
-    appDir,
-    "package.json",
-    JSON.stringify(
-      {
-        name: "fixture-inspect-app",
-        private: true,
-        type: "module",
-      },
-      null,
-      2,
-    ),
-  );
-  writeProjectFile(
-    appDir,
-    "vite.config.ts",
-    `import { defineConfig } from "vite";
-import { pracht } from ${JSON.stringify(vitePluginImport)};
-
-export default defineConfig({
-  plugins: [pracht()],
-  resolve: {
-    alias: {
-      "@pracht/adapter-node": ${JSON.stringify(nodeAdapterImportPath)},
-      "@pracht/core": ${JSON.stringify(coreImportPath)},
-    },
-  },
-});
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/routes.ts",
-    `import { defineApp, group, route, timeRevalidate } from "@pracht/core";
-
-export const app = defineApp({
-  shells: {
-    app: () => import("./shells/app.tsx"),
-  },
-  middleware: {
-    auth: () => import("./middleware/auth.ts"),
-  },
-  routes: [
-    group({ shell: "app", middleware: ["auth"] }, [
-      route("/dashboard", {
-        component: () => import("./routes/dashboard.tsx"),
-        loader: () => import("./server/dashboard-loader.ts"),
-        hydration: "full",
-        prefetch: "hover",
-        render: "isg",
-        revalidate: timeRevalidate(60),
-        speculation: "prefetch",
-      }),
-    ]),
-  ],
-});
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/routes/dashboard.tsx",
-    `import type { RouteComponentProps } from "@pracht/core";
-
-export function Component({ data }: RouteComponentProps) {
-  return <main>{JSON.stringify(data)}</main>;
-}
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/server/dashboard-loader.ts",
-    `import type { LoaderArgs } from "@pracht/core";
-
-export async function loader(_args: LoaderArgs) {
-  return { ok: true };
-}
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/shells/app.tsx",
-    `import type { ShellProps } from "@pracht/core";
-
-export function Shell({ children }: ShellProps) {
-  return <div>{children}</div>;
-}
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/middleware/auth.ts",
-    `import type { MiddlewareFn } from "@pracht/core";
-
-export const middleware: MiddlewareFn = async (_args, next) => {
-  return next();
-};
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/api/health.ts",
-    `import type { BaseRouteArgs } from "@pracht/core";
-
-export function GET(_args: BaseRouteArgs) {
-  return Response.json({ ok: true });
-}
-
-export async function POST(_args: BaseRouteArgs) {
-  return Response.json({ created: true }, { status: 201 });
-}
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "src/api/webhook.ts",
-    `import type { BaseRouteArgs } from "@pracht/core";
-
-export default async function handler(_args: BaseRouteArgs) {
-  return Response.json({ received: true });
-}
-`,
-  );
-  writeProjectFile(
-    appDir,
-    "dist/client/.vite/manifest.json",
-    JSON.stringify(
-      {
-        "virtual:pracht/client": {
-          file: "assets/client.js",
-          imports: ["assets/vendor.js"],
-        },
-        "src/routes/dashboard.tsx": {
-          css: ["assets/dashboard.css"],
-          file: "assets/dashboard.js",
-          imports: ["assets/vendor.js"],
-          src: "src/routes/dashboard.tsx",
-        },
-        "src/shells/app.tsx": {
-          css: ["assets/app.css"],
-          file: "assets/app.js",
-          imports: ["assets/vendor.js"],
-          src: "src/shells/app.tsx",
-        },
-        "assets/vendor.js": {
-          file: "assets/vendor.js",
-        },
-      },
-      null,
-      2,
-    ),
-  );
-}
-
-function writePagesApp(appDir) {
-  writeProjectFile(
-    appDir,
-    "package.json",
-    JSON.stringify(
-      {
-        name: "fixture-pages-app",
-        private: true,
-        dependencies: {
-          "@pracht/adapter-node": "workspace:*",
-          "@pracht/cli": "workspace:*",
-        },
-      },
-      null,
-      2,
-    ),
-  );
-  writeProjectFile(
-    appDir,
-    "vite.config.ts",
-    `import { defineConfig } from "vite";
-import { pracht } from "@pracht/vite-plugin";
-
-export default defineConfig({
-  plugins: [pracht({ pagesDir: "/src/pages" })],
-});
-`,
-  );
-}
-
-function writeProjectFile(appDir, relativePath, contents) {
-  const filePath = resolve(appDir, relativePath);
-  mkdirSync(dirname(filePath), { recursive: true });
-  writeFileSync(filePath, contents.endsWith("\n") ? contents : `${contents}\n`, "utf-8");
-}
